requirements.txt
opencv-python
pandas
PyInstaller
eel
Pillow
pynput
openpyxl
google-api-python-client